changes types

+++ includes/libft.h
@@ -91,7 +91,7 @@ void				ft_lstadd(t_list **alst, t_list *new);
 void				ft_lstiter(t_list *lst, void (*f)(t_list *elem));
 t_list				*ft_lstmap(t_list *lst, t_list *(*f)(t_list *elem));
 
-int					ft_num_len(int n);
+size_t			ft_num_len(int n);
 int					ft_iswhitespace(char c);
 char				*ft_strrev(char *str);
